Upload initial revisions.
This is a first stab at a web-based chat client.  It's also my first public release of any software.
This is basically a simple web-based chat client I intend to continue to play with, evolve, and learn from. 
A large portion of this project has leveraged AI assistance to increase productivity and teach me java!
See the README for more details.